+++ CHANGES.md
@@ -1,3 +1,8 @@
+### Unreleased
+
+- Add `Log.src` for logging (by [xvw](https://xvw.lol))
+- Add template chain (by [xvw](https://xvw.lol))
+
 ### v2.3.0 2025-05-25 Nantes (France)
 
 #### Yocaml